script/requirements.txt
pymongo==3.12.1
python-dotenv==0.19.2
